requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy>=2.0,<3
psycopg[binary]>=3.1
alembic>=1.13
python-dotenv>=1.0
